Added Cargo service

+++ LogisticsServiceWeb/LogisticsService.Core/DbModels/Cargo.cs
@@ -11,12 +11,12 @@ namespace LogisticsService.Core.DbModels
         public int CargoId { get; set; }
         public string? Name { get; set; }
         // Weight saved in kilograms
-        public int  Weight { get; set; }
+        public double  Weight { get; set; }
 
         // Length, Width, Height saved in centimeters
-        public int Length { get; set; }
-        public int Width { get; set; }
-        public int Height { get; set; }
+        public double Length { get; set; }
+        public double Width { get; set; }
+        public double Height { get; set; }
 
         public string? Description { get; set; }
     }
